get started of algorithm search configuration

+++ handlers/message_editor.py
@@ -1,3 +1,4 @@
+from typing import Set
 
 from aiogram.types import Message, chat, CallbackQuery
 
@@ -7,7 +8,9 @@ from utils.Lexicon import LEXICON
 
 class TravelEditor:
     @staticmethod
-    async def edit_message(lexicon_key: str, request):
+    async def edit_message(lexicon_key: str, request, delete_mode=False,
+                           button_texts: Set[str] = None, callback_sign: str = None):
+        '''Высылальщик сообщений, '''
         lexicon_part = LEXICON[lexicon_key]
         redis_key = str(request.from_user.id) + ':last_message'
         last_message_id = await redis_data.get_data(redis_key)
@@ -18,13 +21,21 @@ class TravelEditor:
             message_object = request
 
         chat_object = message_object.chat
+        print(button_texts)
+        if button_texts:
+            keyboard = await InlineCreator.create_markup(lexicon_part,
+                                                         button_texts=button_texts, callback_sign=callback_sign)
+        else:
+            keyboard = await InlineCreator.create_markup(lexicon_part)
 
-        keyboard = await InlineCreator.create_markup(lexicon_part)
         message_text = lexicon_part['message_text']
-        await chat.Chat.delete_message(self=chat_object, message_id=last_message_id)
-        new_message = await message_object.answer(text=message_text, reply_markup=keyboard)
 
-        await redis_data.set_data(redis_key, new_message.message_id)
+        if delete_mode:
+            await chat.Chat.delete_message(self=chat_object, message_id=last_message_id)
+            new_message = await message_object.answer(text=message_text, reply_markup=keyboard)
+            await redis_data.set_data(redis_key, new_message.message_id)
+        else:
+            await message_object.edit_text(text=message_text, reply_markup=keyboard)
 
 
 travel_editor = TravelEditor()
